refactor: Move metrics metadata into package "internal"
