+++ packages/cosmic-swingset/test/snapshots/prometheus.test.js.md
@@ -15,6 +15,7 @@ Generated by [AVA](https://avajs.dev).
     # TYPE cosmic_swingset_inbound_actions_total counter␊
     cosmic_swingset_inbound_actions_total{actionType="CORE_EVAL"} %f %@␊
     # HELP swingset_vat_startup Vat startup time (ms)␊
+    # UNIT swingset_vat_startup ms␊
     # TYPE swingset_vat_startup histogram␊
     swingset_vat_startup_count %f %@␊
     swingset_vat_startup_sum %f %@␊
@@ -160,6 +161,7 @@ Generated by [AVA](https://avajs.dev).
     # TYPE memoryUsage_arrayBuffers gauge␊
     memoryUsage_arrayBuffers %f %@␊
     # HELP swingset_crank_processing_time Processing time per crank (ms)␊
+    # UNIT swingset_crank_processing_time ms␊
     # TYPE swingset_crank_processing_time histogram␊
     swingset_crank_processing_time_count %f %@␊
     swingset_crank_processing_time_sum %f %@␊
@@ -175,6 +177,7 @@ Generated by [AVA](https://avajs.dev).
     swingset_crank_processing_time_bucket{le="91"} %f %@␊
     swingset_crank_processing_time_bucket{le="+Inf"} %f %@␊
     # HELP swingset_block_processing_seconds Processing time per block␊
+    # UNIT swingset_block_processing_seconds s␊
     # TYPE swingset_block_processing_seconds histogram␊
     swingset_block_processing_seconds_count %f %@␊
     swingset_block_processing_seconds_sum %f %@␊
