+++ packages/internal/test/snapshots/exports.test.js.md
@@ -9,11 +9,17 @@ Generated by [AVA](https://avajs.dev).
 > Snapshot 1
 
     [
+      'BLOCK_HISTOGRAM_METRICS',
       'BridgeBigIntShape',
       'BridgeId',
       'CosmosInitKeyToBridgeId',
+      'HISTOGRAM_METRICS',
+      'KERNEL_STATS_METRICS',
+      'KERNEL_STATS_SUM_METRICS',
+      'KERNEL_STATS_UPDOWN_METRICS',
       'NonNullish',
       'PromiseAllOrErrors',
+      'QueueMetricAspect',
       'StorageNodeShape',
       'TRUE',
       'UnguardedHelperI',
@@ -37,6 +43,8 @@ Generated by [AVA](https://avajs.dev).
       'logLevels',
       'makeLimitedConsole',
       'makeMeasureSeconds',
+      'makeQueueMetrics',
+      'makeQueueMetricsMeta',
       'makeTempDirFactory',
       'makeTracer',
       'mustMatch',
